HelpDesk_Application/BE/API/requirements.txt
anyio==3.5.0
asgiref==3.5.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
fastapi==0.74.1
h11==0.13.0
idna==3.3
PyJWT==2.3.0
numpy==1.22.2
pandas==1.4.1
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.5
openpyxl==3.0.9